refactor: move remaining modules from ui-utils
