--- apps/meteor/app/ui-utils/client/index.ts
@@ -1,3 +1,0 @@
-export { messageBox } from './lib/messageBox';
-export { LegacyRoomManager } from './lib/LegacyRoomManager';
-export { upsertMessage, RoomHistoryManager } from './lib/RoomHistoryManager';
