+++ apps/meteor/client/components/message/hooks/useMarkAsUnreadMutation.ts
@@ -2,7 +2,7 @@ import type { IMessage, ISubscription } from '@rocket.chat/core-typings';
 import { useEndpoint, useToastMessageDispatch } from '@rocket.chat/ui-contexts';
 import { useMutation } from '@tanstack/react-query';
 
-import { LegacyRoomManager } from '../../../../app/ui-utils/client';
+import { LegacyRoomManager } from '../../../lib/LegacyRoomManager';
 
 export const useMarkAsUnreadMutation = () => {
 	const dispatchToastMessage = useToastMessageDispatch();
