+++ apps/meteor/client/hooks/useMessageboxAppsActionButtons.ts
@@ -5,8 +5,8 @@ import { useTranslation } from 'react-i18next';
 import { useAppActionButtons, getIdForActionButton } from './useAppActionButtons';
 import { useApplyButtonFilters } from './useApplyButtonFilters';
 import { UiKitTriggerTimeoutError } from '../../app/ui-message/client/UiKitTriggerTimeoutError';
-import type { MessageBoxAction } from '../../app/ui-utils/client/lib/messageBox';
 import { Utilities } from '../../ee/lib/misc/Utilities';
+import type { MessageBoxAction } from '../lib/messageBox';
 import { useUiKitActionManager } from '../uikit/hooks/useUiKitActionManager';
 
 export const useMessageboxAppsActionButtons = () => {
