+++ apps/meteor/client/importPackages.ts
@@ -1,3 +1,2 @@
 import '../app/lib/client';
 import '../app/utils/client';
-import '../app/ui-utils/client';
