+++ apps/meteor/client/lib/LegacyRoomManager.ts
@@ -4,14 +4,14 @@ import { createPredicateFromFilter } from '@rocket.chat/mongo-adapter';
 import { clientCallbacks } from '@rocket.chat/ui-client';
 import type { Filter } from 'mongodb';
 
+import { RoomManager } from './RoomManager';
+import { Messages, Subscriptions } from '../stores';
 import { upsertMessage, RoomHistoryManager } from './RoomHistoryManager';
-import { RoomManager } from '../../../../client/lib/RoomManager';
-import { roomCoordinator } from '../../../../client/lib/rooms/roomCoordinator';
-import { fireGlobalEvent } from '../../../../client/lib/utils/fireGlobalEvent';
-import { getConfig } from '../../../../client/lib/utils/getConfig';
-import { modifyMessageOnFilesDelete } from '../../../../client/lib/utils/modifyMessageOnFilesDelete';
-import { Messages, Subscriptions } from '../../../../client/stores';
-import { sdk } from '../../../utils/client/lib/SDKClient';
+import { roomCoordinator } from './rooms/roomCoordinator';
+import { fireGlobalEvent } from './utils/fireGlobalEvent';
+import { getConfig } from './utils/getConfig';
+import { modifyMessageOnFilesDelete } from './utils/modifyMessageOnFilesDelete';
+import { sdk } from '../../app/utils/client/lib/SDKClient';
 
 const maxRoomsOpen = parseInt(getConfig('maxRoomsOpen') ?? '5') || 5;
