+++ apps/meteor/client/lib/messageBox.ts
@@ -2,7 +2,7 @@ import type { IMessage, IRoom } from '@rocket.chat/core-typings';
 import type { Keys as IconName } from '@rocket.chat/icons';
 import type { TranslationKey } from '@rocket.chat/ui-contexts';
 
-import type { ChatAPI } from '../../../../client/lib/chats/ChatAPI';
+import type { ChatAPI } from './chats/ChatAPI';
 
 export type MessageBoxAction = {
 	label: TranslationKey;
