+++ apps/meteor/client/views/navigation/sidepanel/hooks/useRoomMenuActions.spec.ts
@@ -14,7 +14,7 @@ jest.mock('../../../../../client/lib/rooms/roomCoordinator', () => ({
 	},
 }));
 
-jest.mock('../../../../../app/ui-utils/client', () => ({
+jest.mock('../../../../../client/lib/LegacyRoomManager', () => ({
 	LegacyRoomManager: {
 		close: jest.fn(),
 	},
