+++ apps/meteor/client/views/room/providers/RoomProvider.tsx
@@ -8,10 +8,10 @@ import UserCardProvider from './UserCardProvider';
 import { useRedirectOnSettingsChanged } from './hooks/useRedirectOnSettingsChanged';
 import { useUsersNameChanged } from './hooks/useUsersNameChanged';
 import { UserAction } from '../../../../app/ui/client/lib/UserAction';
-import { RoomHistoryManager, useRoomHistoryState } from '../../../../app/ui-utils/client/lib/RoomHistoryManager';
 import { omit } from '../../../../lib/utils/omit';
 import { useFireGlobalEvent } from '../../../hooks/useFireGlobalEvent';
 import { useRoomRolesQuery } from '../../../hooks/useRoomRolesQuery';
+import { RoomHistoryManager, useRoomHistoryState } from '../../../lib/RoomHistoryManager';
 import { RoomManager } from '../../../lib/RoomManager';
 import { roomCoordinator } from '../../../lib/rooms/roomCoordinator';
 import ImageGalleryProvider from '../../../providers/ImageGalleryProvider';
