+++ apps/meteor/client/views/room/providers/hooks/useRedirectOnSettingsChanged.ts
@@ -2,7 +2,7 @@ import type { ISubscription } from '@rocket.chat/core-typings';
 import { useRouter } from '@rocket.chat/ui-contexts';
 import { useEffect } from 'react';
 
-import { LegacyRoomManager } from '../../../../../app/ui-utils/client';
+import { LegacyRoomManager } from '../../../../lib/LegacyRoomManager';
 import { roomCoordinator } from '../../../../lib/rooms/roomCoordinator';
 
 const routeNameToRoomTypeMap: Record<string, string> = {
